Initial AMD Profiling (#339)
* rocprof: implement ROCm profiling

This uses rocPROF to fetch some interesting data and put it
in the profile_data directory, the download link of which
is then returned to the user.

* rocprof: post-process rocprof results

rocPROF generates one trace for every process. Simply combine them
together into a single trace for ease of use. Also remove the
individual traces are they are no longer useful afterwards.

* rocprof: also output code objects